Add Support for SoundCloud Fix bug that prevented Youtube embeds from working if start time provided Updated license classto provide better messages on activation

+++ includes/ps/lts-license-handler.php
@@ -16,7 +16,8 @@ function lts_plugin_updater() {
             'version' 	=> LINK_TIMESTAMP_VERSION, 				// current version number
             'license' 	=> $license_key, 		// license key (used get_option above to retrieve from DB)
             'item_name' => LINK_TIMESTAMP_EDD_NAME, 	// name of this plugin
-            'author' 	=> 'Arelthia Phillips'  // author of this plugin
+            'author' 	=> 'Arelthia Phillips',  // author of this plugin
+            'beta'      => LINK_TIMESTAMP_BETA
         )
     );
 
@@ -134,17 +135,82 @@ function lts_activate_license() {
         $response = wp_remote_post( LINK_TIMESTAMP_EDD_STORE_URL, array( 'timeout' => 15, 'sslverify' => false, 'body' => $api_params ) );
 
         // make sure the response came back okay
-        if ( is_wp_error( $response ) ){
-            return false;
-        }
+        if ( is_wp_error( $response ) || 200 !== wp_remote_retrieve_response_code( $response ) ) {
+
+            if ( is_wp_error( $response ) ) {
+                $message = $response->get_error_message();
+            } else {
+                $message = __( 'An error occurred, please try again.' );
+            }
+
+        } else {
+
 
         // decode the license data
         $license_data = json_decode( wp_remote_retrieve_body( $response ) );
 
+        if ( false === $license_data->success ) {
+
+                switch( $license_data->error ) {
+
+                    case 'expired' :
+
+                        $message = sprintf(
+                            __( 'Your license key expired on %s.' ),
+                            date_i18n( get_option( 'date_format' ), strtotime( $license_data->expires, current_time( 'timestamp' ) ) )
+                        );
+                        break;
+
+                    case 'revoked' :
+
+                        $message = __( 'Your license key has been disabled.' );
+                        break;
+
+                    case 'missing' :
+
+                        $message = __( 'Invalid license.' );
+                        break;
+
+                    case 'invalid' :
+                    case 'site_inactive' :
+
+                        $message = __( 'Your license is not active for this URL.' );
+                        break;
+
+                    case 'item_name_mismatch' :
+
+                        $message = sprintf( __( 'This appears to be an invalid license key for %s.' ), LINK_TIMESTAMP_EDD_NAME );
+                        break;
+
+                    case 'no_activations_left':
+
+                        $message = __( 'Your license key has reached its activation limit.' );
+                        break;
+
+                    default :
+
+                        $message = __( 'An error occurred, please try again.' );
+                        break;
+                }
+            }        
+        }
+
+            // Check if anything passed on a message constituting a failure
+        if ( ! empty( $message ) ) {
+            $base_url = admin_url( LINK_TIMESTAMP_LICENSE_PAGE );
+            $redirect = add_query_arg( array( 'sl_activation' => 'false', 'message' => urlencode( $message ) ), $base_url );
+
+            wp_redirect( $redirect );
+            exit();
+        }
+
+
+
         // $license_data->license will be either "valid" or "invalid"
 
         update_option( 'lts_license_status', $license_data->license );
-
+        wp_redirect( admin_url( LINK_TIMESTAMP_LICENSE_PAGE ) );
+        exit();
     }
 }
 add_action('admin_init', 'lts_activate_license');
@@ -177,17 +243,32 @@ function lts_deactivate_license() {
         $response = wp_remote_post( LINK_TIMESTAMP_EDD_STORE_URL, array( 'timeout' => 15, 'sslverify' => false, 'body' => $api_params ) );
 
         // make sure the response came back okay
-        if ( is_wp_error( $response ) )
-            return false;
+        if ( is_wp_error( $response ) || 200 !== wp_remote_retrieve_response_code( $response ) ) {
+
+            if ( is_wp_error( $response ) ) {
+                $message = $response->get_error_message();
+            } else {
+                $message = __( 'An error occurred, please try again.' );
+            }
+
+            $base_url = admin_url( LINK_TIMESTAMP_LICENSE_PAGE );
+            $redirect = add_query_arg( array( 'sl_activation' => 'false', 'message' => urlencode( $message ) ), $base_url );
+
+            wp_redirect( $redirect );
+            exit();
+        }
 
         // decode the license data
         $license_data = json_decode( wp_remote_retrieve_body( $response ) );
 
         // $license_data->license will be either "deactivated" or "failed"
-        if( $license_data->license == 'deactivated' )
+        if( $license_data->license == 'deactivated' ){
             delete_option( 'lts_license_status' );
+        }
 
-    }
+        wp_redirect( admin_url( LINK_TIMESTAMP_LICENSE_PAGE ) );
+        exit();
+    }    
 }
 add_action('admin_init', 'lts_deactivate_license');
 /************************************
@@ -226,3 +307,37 @@ function lts_check_license() {
         // this license is no longer valid
     }
 }
+
+
+
+/**
+ * This is a means of catching errors from the activation method above and displaying it to the customer
+ */
+function lts_admin_notices() {
+    if ( isset( $_GET['sl_activation'] ) && ! empty( $_GET['message'] ) ) {
+
+        switch( $_GET['sl_activation'] ) {
+
+            case 'false':
+                $message = urldecode( $_GET['message'] );
+                ?>
+                <div class="error">
+                    <p><?php echo $message; ?></p>
+                </div>
+                <?php
+                break;
+
+            case 'true':
+            default:
+                $message = 'License activated successfully.';
+                ?>
+                <div class="success">
+                    <p><?php echo $message; ?></p>
+                </div>
+                <?php
+                break;
+
+        }
+    }
+}
+add_action( 'admin_notices', 'lts_admin_notices' );
